Move storage test object to fixture
